pyproject.toml
[project]
name = "custom-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-community>=0.3.5",
    "langchain-groq>=0.2.1",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.2.0",
    "langchain>=0.3.7",
    "llama-cpp-python>=0.3.2",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.2.1",
    "streamlit>=1.40.0",
    "torch==2.3.1",
    "torchvision>=0.18.1",
    "wurlitzer>=3.1.1",
]
